lab12_1 completed

+++ lab12_1.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 
-int fibonacci(int);
+long long fibonacci(int);
 
 int main(){
     int x;
@@ -10,7 +10,7 @@ int main(){
     return 0; 
 }
 
-int fibonacci(int n){
+long long fibonacci(int n){
     if(n == 0) return 0;
     else if (n == 1) return 1;
     else return fibonacci(n-1) + fibonacci(n-2);
